requirements.txt
Flask==2.1.0
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
reportlab==3.6.9
python-dotenv==0.20.0